eletivaspro/requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.3
django-easy-pdf==0.1.1
django-extensions==3.0.7
django-filter==2.3.0
django-import-export==1.2.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.1
et-xmlfile==1.0.1
gunicorn==19.9.0
html5lib==1.0.1
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.1.0
psycopg2-binary==2.8.3
PyPDF2==1.26.0
pytz==2019.1
PyYAML==5.1.1
reportlab==3.5.23
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
webencodings==0.5.1
whitenoise==4.1.3
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
